Refactor data imports: move navbarData and servicesData to new data directory
