+++ data/navbarData.ts
@@ -1,0 +1,7 @@
+export const navbarData = [
+  { title: "Home", href: "/" },
+  // { title: "Services", href: "/services" },
+  { title: "Resume", href: "/resume" },
+  { title: "Work", href: "/work" },
+  { title: "Contact", href: "/contact" },
+];
