+++ data/servicesData.ts
@@ -1,11 +1,3 @@
-export const navbarData = [
-  { title: "Home", href: "/" },
-  // { title: "Services", href: "/services" },
-  { title: "Resume", href: "/resume" },
-  { title: "Work", href: "/work" },
-  { title: "Contact", href: "/contact" },
-];
-
 export const servicesData = [
   {
     _id: "01",
